kill code_model_impl
