--- crates/ra_hir/src/code_model_impl.rs
@@ -1,2 +1,0 @@
-mod konst; // `const` is invalid ident :(
-pub(crate) mod function;
